+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   `org.datalevin:datalevin-java:<version>` to maven central.
 - [Lib] Release Datalevin as a Python library `datalevin`to PyPI.
 - [Lib] Release Datalevin as a Node.js library `datalevin-node` to npm.
+- [Lib] Release a trimmed embedded-only jar
+  `datalevin/datalevin-embedded:<version>` to Clojars.
 - [API] JSON API. [#182](https://github.com/datalevin/datalevin/issues/182)
 - [MCP] Built-in stdio based MCP server. Details in [doc](doc/mcp.md).
 - [Datalog] allow to register a `:db/udf` to specify a user defined function.